requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pillow==10.4.0
numpy==1.26.4

# Real-ESRGAN dependencies
realesrgan==0.3.0
basicsr==1.4.2
facexlib==0.3.0
gfpgan==1.3.8
opencv-python-headless==4.10.0.84

# PyTorch CPU wheels (Railway Free tidak ada GPU)
torch==2.5.1
torchvision==0.20.1
